app/requirements.txt
aiodataloader==0.3.0
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.8.1
aniso8601==9.0.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
bcrypt==4.0.1
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
ecdsa==0.18.0
fastapi==0.88.0
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.57.0
graphene==3.1.1
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==2.0.1
h11==0.14.0
httplib2==0.21.0
httptools==0.5.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.1
multidict==6.0.3
oauthlib==3.2.2
passlib==1.7.4
promise==2.3
protobuf==4.21.12
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.22.0
starlette-graphene3==0.6.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
